renamed `SlickTestUtil` to `SlickExplainUtil`
